Fixed some issues with the powershell scripts inside of .vscode
I had only tested the other configuration on a single project and in practice saw that the current vscode terminal location would cause a fail if it was in use. Now the build task uses a reference location...

Also fixed a few of the build tasks which seemed to work before, but werent calling anymore for some reason using the & modifier. Now thye seem to always work using the powershell command to open the ps1 files.